% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topography_service.R
\name{topography_source}
\alias{topography_source}
\title{File path for topography}
\usage{
topography_source(x = c("gebco", "aws"))
}
\arguments{
\item{x}{name of the 'label' of the topography source, see \link{topography_services}}
}
\value{
file path to an installed .xml configuration of an online raster (WMS) tile service
}
\description{
Obtain the installed file path for a GDAL WMS source for an online topography.
}
\examples{
topography_source("aws")
topography_source("gebco")
}
